JWT: add jwt to pom

+++ server/src/main/java/pl/justdrinkjava/JustDrinkJava/mapper/UserMapper.java
@@ -14,7 +14,7 @@ public class UserMapper {
         }
         
         return UserDTO.builder()
-                .id(user.getId())
+                .id(user.getId().intValue())
                 .username(user.getUsername())
                 .email(user.getEmail())
                 .createdAt(user.getCreatedAt())
